+++ src/main/java/com/jpaweekily/domain/order/application/OrderService.java
@@ -1,0 +1,4 @@
+package com.jpaweekily.domain.order.application;
+
+public interface OrderService {
+}
