+++ src/main/java/com/jpaweekily/domain/order/infrastructrue/OrderRepository.java
@@ -1,4 +1,4 @@
-package com.jpaweekily.domain.order.repository;
+package com.jpaweekily.domain.order.infrastructrue;
 
 import com.jpaweekily.domain.order.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
