--- src/main/java/com/jpaweekily/domain/order/service/OrderService.java
@@ -1,4 +1,0 @@
-package com.jpaweekily.domain.order.service;
-
-public interface OrderService {
-}
