+++ src/main/java/com/jpaweekily/domain/product/application/ProductService.java
@@ -1,0 +1,4 @@
+package com.jpaweekily.domain.product.application;
+
+public interface ProductService {
+}
