--- src/main/java/com/jpaweekily/domain/product/service/ProductService.java
@@ -1,4 +1,0 @@
-package com.jpaweekily.domain.product.service;
-
-public interface ProductService {
-}
