+++ src/main/java/com/jpaweekily/domain/user/application/UserService.java
@@ -1,0 +1,4 @@
+package com.jpaweekily.domain.user.application;
+
+public interface UserService {
+}
