--- src/main/java/com/jpaweekily/domain/user/service/UserService.java
@@ -1,4 +1,0 @@
-package com.jpaweekily.domain.user.service;
-
-public interface UserService {
-}
